[db-slick] Deprecate db-slick-psql... move everything into db-slick
